Break up app into components
