--- js-pkg/create-y-sweet-app/src/templates/nextjs/.env.default
@@ -1,1 +1,0 @@
-CONNECTION_STRING="y-sweet-connection-string-goes-here"
