+++ js-pkg/create-y-sweet-app/src/templates/nextjs/.gitignore
@@ -1,1 +1,4 @@
 .next
+next-env.d.ts
+.env
+.env.*
