+++ js-pkg/create-y-sweet-app/src/templates/remix/.gitignore
@@ -1,0 +1,2 @@
+.env
+.env.*
